Add files via upload
This script takes in an input from Google's Vision API and runs it against our list of sub-strings of expected trash names. It returns 0 if it is trash, 1 if it is recycling, or 2 if it is compost. It returns 0 otherwise for edge cases such as unidentified objects and when the ultrasonic sensor detects a hand being placed in front of it.